Batch_Inference/export/code/requirements.txt
Pillow
tensorflow==2.3.1